% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaKnockoff.R
\name{copula_knockoff}
\alias{copula_knockoff}
\title{Knockoff Construction through Copula}
\usage{
copula_knockoff(data.x, R = NULL, marginal_list = NULL)
}
\arguments{
\item{data.x}{input data}

\item{R}{correlation matrix, can be provided, default value is NULL, where the algorithm will estimate it}

\item{marginal_list}{marginal list, containing mean and standard deviation}
}
\value{
A list of correlation and constructed knockoff
}
\description{
Construct knockoffs by estimating Gaussian Copulas first
}
\examples{
out_copula = gaussian_copula(p = 100, n = 500, type = "diag")
out_data = normal_distribution_generate(copula = out_copula, p = 100, n = 500)
out = copula_knockoff(out_data$data.normal)

}
\keyword{copula,}
\keyword{knockoff}
