% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaKnockoff.R
\name{copulaKnockoff}
\alias{copulaKnockoff}
\title{Knockoff Construction through Copula}
\usage{
copulaKnockoff(data.x, R = NULL, marginal_list = NULL)
}
\arguments{
\item{data.x}{input data}

\item{R}{correlation matrix, can be provided, default value is NULL, where the algorithm will estimate it}

\item{marginal_list}{marginal list, containing mean and standard deviation}
}
\value{
A list of correlation and constructed knockoff
}
\description{
Construct knockoffs by estimating Gaussian Copulas first
}
\examples{
out_copula = gaussianCopula(p = 100, n = 500, type = "diag")
out_data = gaussianDistributionGeneration(copula = out_copula, p = 100, n = 500)
out = copulaKnockoff(out_data$data)

}
\keyword{copula,}
\keyword{knockoff}
