% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaDistributionGeneration.R
\name{gamma_distribution_generate}
\alias{gamma_distribution_generate}
\title{Gamma Distribution Generation with Copula}
\usage{
gamma_distribution_generate(copula, p, n)
}
\arguments{
\item{copula}{Copula used to generate gamma distribution}

\item{p}{Number of features}

\item{n}{Number of samples}
}
\value{
A list of data, multivariate mean and multivariate standard deviation
}
\description{
Generate gamma distribution with provided copula of p features and n samples
}
\examples{

out_copula = gaussian_copula(p = 100, n = 500, type = "diag")
out_data = gamma_distribution_generate(copula = out_copula, p = 100, n = 500)

}
\keyword{copula,}
\keyword{distribution}
\keyword{gamma}
