% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianCopula.R
\name{gaussianCopula}
\alias{gaussianCopula}
\title{Gaussian Copula Generation}
\usage{
gaussianCopula(p, n, type = "diag")
}
\arguments{
\item{p}{Dimension of the covariance}

\item{n}{Number of samples}

\item{type}{Structure of the covariance matrix, has option "diag", "toeplitz" and "unstructured"}
}
\value{
A list of correlation anddata
}
\description{
Generate a Gaussian copula with specific covariance structures
}
\examples{
out = gaussianCopula(1000, 100, "diag")


}
\keyword{Copula}
\keyword{Gaussian,}
