% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianDistribution.R
\name{gaussianDistribution}
\alias{gaussianDistribution}
\title{Gaussian Distribution Generation}
\usage{
gaussianDistribution(p, n, diag_cov = "unstructure")
}
\arguments{
\item{p}{Dimension of the covariance}

\item{n}{Number of samples}

\item{diag_cov}{Structure of the covariance matrix}
}
\value{
A list contains correlation, simulated data, multivariate mean, multivariate standard deviation, covariance
}
\description{
Generate a Gaussian Distribution with specific covariance
}
\examples{
out = gaussianDistribution(100, 50, "unstructure")
out = gaussianDistribution(100, 50, "diagonal")
out = gaussianDistribution(100, 50, "toeplitz")

}
\keyword{covariance,}
\keyword{gaussian}
