% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianDistributionGeneration.R
\name{gaussianDistributionGeneration}
\alias{gaussianDistributionGeneration}
\title{Gaussian Distribution Generation with Copula}
\usage{
gaussianDistributionGeneration(copula, p, n)
}
\arguments{
\item{copula}{Copula used to generate normal distribution}

\item{p}{Number of features}

\item{n}{Number of samples}
}
\value{
A list of data, multivariate mean and multivariate standard deviation
}
\description{
Generate gaussian distribution with provided copula of p features and n samples
}
\examples{

out_copula = gaussianCopula(p = 100, n = 500, type = "diag")
out_data = gaussianDistributionGeneration(copula = out_copula, p = 100, n = 500)

}
\keyword{copula,}
\keyword{distribution}
\keyword{gaussian}
