% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoGenerationModel.R
\name{Lasso_generation_model}
\alias{Lasso_generation_model}
\title{Data Generation with Lasso Model}
\usage{
Lasso_generation_model(data_signal, p_freq)
}
\arguments{
\item{data_signal}{The covariate used to generate the data}

\item{p_freq}{The percentage of covariates that are set to be 0}
}
\value{
A list containing the data, covariant, the mask and p_freq
}
\description{
Generate data with the Lasso Model
}
\examples{
out_X = gaussian_distribution(100, 50, "unstructure")
out_Y_Lasso = Lasso_generation_model(out_X$X, 0.5)

}
\keyword{generation}
\keyword{lasso,}
