% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoFDR.R
\name{lassoFDR}
\alias{lassoFDR}
\title{Knockoff Variable Selection through Lasso Regression while Controlling the FDR}
\usage{
lassoFDR(x, x.knockoff, y, mask = NULL, family, fdr_rate = 0.1)
}
\arguments{
\item{x}{Original data}

\item{x.knockoff}{knockoff copy of the original data}

\item{y}{Response variable}

\item{mask}{The variables that are masked from the purpose}

\item{family}{what kind of}

\item{fdr_rate}{Controlling the False Discovery Rate of the knockoff}
}
\description{
Select variable based on knockoff constructed from lasso Regression
}
\examples{
out_copula = gaussianCopula(p = 100, n = 500, type = "diag")
out_data = gaussianDistributionGeneration(copula = out_copula, p = 100, n = 500)
out = copulaKnockoff(out_data$data)
lasso_out = LassoGenerationModel(data_signal = out_data$data, p_freq = 0.3)
lassoFDR(x = out_data$data, x.knockoff = out$x.knockoff.copy, y = lasso_out$y, family = "gaussian")

}
\keyword{FDR}
\keyword{Lasso,}
\keyword{knockoff,}
