% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalDistributionGeneration.R
\name{normal_distribution_generate}
\alias{normal_distribution_generate}
\title{Normal Distribution Generation with Copula}
\usage{
normal_distribution_generate(copula, p, n)
}
\arguments{
\item{copula}{Copula used to generate normal distribution}

\item{p}{Number of features}

\item{n}{Number of samples}
}
\value{
A list of data, multivariate mean and multivariate standard deviation
}
\description{
Generate normal distribution with provided copula of p features and n samples
}
\examples{

out_copula = gaussian_copula(p = 100, n = 500, type = "diag")
out_data = normal_distribution_generate(copula = out_copula, p = 100, n = 500)

}
\keyword{copula,}
\keyword{distribution}
\keyword{normal}
