% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeFDR.R
\name{ridgeFDR}
\alias{ridgeFDR}
\title{Knockoff Variable Selection through Ridge while Controlling the FDR}
\usage{
ridgeFDR(x, x.knockoff, y, mask = NULL, family = NULL, fdr_rate = 0.1)
}
\arguments{
\item{x}{Original data}

\item{x.knockoff}{knockoff copy of the original data}

\item{y}{Response variable}

\item{mask}{The variables that are masked from the purpose}

\item{family}{what kind of}

\item{fdr_rate}{Controlling the False Discovery Rate of the knockoff}
}
\description{
Select variable based on knockoff constructed from Ridge
}
\examples{
out_copula = gaussian_copula(p = 100, n = 500, type = "diag")
out_data = normal_distribution_generate(copula = out_copula, p = 100, n = 500)
out = copula_knockoff(out_data$data.normal)
lasso_out = Lasso_generation_model(data_signal = out_data$data.normal, p_freq = 0.3)
ridgeFDR(x = out_data$data.normal, x.knockoff = out$x.knockoff.copy, y = lasso_out$y)

}
\keyword{FDR}
\keyword{Ridge,}
\keyword{knockoff,}
