% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tDistributionGeneration.R
\name{t_distribution_generate}
\alias{t_distribution_generate}
\title{t Distribution Generation with Copula}
\usage{
t_distribution_generate(copula, p, n)
}
\arguments{
\item{copula}{Copula used to generate t distribution}

\item{p}{Number of features}

\item{n}{Number of samples}
}
\value{
A list of data, multivariate mean and multivariate standard deviation
}
\description{
Generate t distribution with provided copula of p features and n samples
}
\examples{

out_copula = gaussian_copula(p = 100, n = 500, type = "diag")
out_data = t_distribution_generate(copula = out_copula, p = 100, n = 500)

}
\keyword{copula,}
\keyword{distribution}
\keyword{t}
