% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncatedGaussianKnockoff.R
\name{truncatedGaussianKnockoff}
\alias{truncatedGaussianKnockoff}
\title{Truncated Gaussian knockoff}
\usage{
truncatedGaussianKnockoff(x, Sigma, amp = 1)
}
\arguments{
\item{x}{original data}

\item{Sigma}{covariance matrix of the original X}

\item{amp}{The amplitude of the gaussian copula, default value is set to 1, upper limit is 2.}
}
\value{
knockoff generated
}
\description{
Construct Gaussian knockoff by truncating the diagonal variables
}
\examples{

library(mvtnorm)
p = 20
n = 100
u = eigen(cov(matrix(rnorm(p * n), nrow = n)))$vectors
d = runif(p, 1, 10)
Sigma = u \%*\% diag(d) \%*\% t(u)
x = rmvnorm(n, rep(0, p), Sigma)
truncatedGaussianKnockoff(x, Sigma, 1)
}
\keyword{knockoff}
\keyword{truncated,}
